Convert from doubles to singles

+++ Devcade2048/App/Config.cs
@@ -4,7 +4,7 @@ namespace Devcade2048.App;
 
 public static class Config {
     public static int StartTiles = 2;
-    public static double FourChance = 0.1d;
+    public static float FourChance = 0.1F;
 
     public static bool LoadGame = false;
 }
