+++ Devcade2048/App/Render/Animation/ToTabAnimationState.cs
@@ -27,7 +27,7 @@ public class ToTabAnimationState : TransientAnimationState {
         // This backgroundColor pixel is stretched to cover everything. 
         Rectangle screen = new Rectangle(0, 0, gd.DisplayMode.Width, gd.DisplayMode.Height);
 
-        DrawAsset(uiCover, screen, Color.White * (float) (1 - FastEnd()));
+        DrawAsset(uiCover, screen, Color.White * (1 - FastEnd()));
         base.Draw();
     }
 }
